docs(spacing): update structure and move to helpers
