+++ .storybook/main.js
@@ -101,9 +101,12 @@ module.exports = {
     '../projects/canopy/src/lib/show-at/docs/show-at.stories.ts',
     '../projects/canopy/src/lib/skeleton/docs/guide.stories.mdx',
     '../projects/canopy/src/lib/skeleton/docs/skeleton.stories.ts',
-    '../projects/canopy/src/lib/spacing/margin/margin.stories.ts',
-    '../projects/canopy/src/lib/spacing/padding/padding.stories.ts',
-    '../projects/canopy/src/lib/spacing/row-gap/row-gap.stories.ts',
+    '../projects/canopy/src/lib/spacing/margin/docs/guide.stories.mdx',
+    '../projects/canopy/src/lib/spacing/margin/docs/margin.stories.ts',
+    '../projects/canopy/src/lib/spacing/padding/docs/guide.stories.mdx',
+    '../projects/canopy/src/lib/spacing/padding/docs/padding.stories.ts',
+    '../projects/canopy/src/lib/spacing/row-gap/docs/guide.stories.mdx',
+    '../projects/canopy/src/lib/spacing/row-gap/docs/row-gap.stories.ts',
     '../projects/canopy/src/lib/sr-alert-message/sr-alert-message.stories.ts',
     '../projects/canopy/src/lib/variant/variant.stories.ts',
     //  Pipes
